match the data type in UpSample and Downsample module

+++ vllm/model_executor/models/qwen2_5_omni_token2wav.py
@@ -751,10 +751,10 @@ class UpSample1d(nn.Module):
     def forward(self, hidden_states):
         channels = hidden_states.shape[1]
 
-        hidden_states = F.pad(hidden_states, (self.pad, self.pad), mode="replicate")
+        hidden_states = F.pad(hidden_states, (self.pad, self.pad), mode="replicate").to(self.filter.dtype)
         hidden_states = self.ratio * F.conv_transpose1d(
             hidden_states, self.filter.expand(channels, -1, -1), stride=self.stride, groups=channels
-        )
+        ).to(self.dtype)
         hidden_states = hidden_states[..., self.pad_left : -self.pad_right]
 
         return hidden_states
@@ -780,8 +780,8 @@ class DownSample1d(nn.Module):
 
     def forward(self, hidden_states):
         channels = hidden_states.shape[1]
-        hidden_states = F.pad(hidden_states, (self.pad_left, self.pad_right), mode="replicate")
-        out = F.conv1d(hidden_states, self.filter.expand(channels, -1, -1), stride=self.stride, groups=channels)
+        hidden_states = F.pad(hidden_states, (self.pad_left, self.pad_right), mode="replicate").to(self.filter.dtype)
+        out = F.conv1d(hidden_states, self.filter.expand(channels, -1, -1), stride=self.stride, groups=channels).to(self.dtype)
         return out
